Split write operations from IBucketReader to separate interface
