+++ DLCS.Repository.Tests/Storage/S3/BucketReaderTests.cs
@@ -20,14 +20,14 @@ namespace DLCS.Repository.Tests.Storage.S3
     public class BucketReaderTests
     {
         private readonly IAmazonS3 s3Client;
-        private readonly ILogger<BucketReader> logger;
-        private readonly BucketReader sut;
+        private readonly ILogger<S3BucketReader> logger;
+        private readonly S3BucketReader sut;
         
         public BucketReaderTests()
         {
             s3Client = A.Fake<IAmazonS3>();
-            logger = A.Fake<ILogger<BucketReader>>();
-            sut = new BucketReader(s3Client, logger);
+            logger = A.Fake<ILogger<S3BucketReader>>();
+            sut = new S3BucketReader(s3Client, logger);
         }
 
         [Fact]
