+++ Orchestrator/Startup.cs
@@ -64,7 +64,8 @@ namespace Orchestrator
                 .AddSingleton<ImageRequestHandler>()
                 .AddSingleton<TimeBasedRequestHandler>()
                 .AddAWSService<IAmazonS3>()
-                .AddSingleton<IBucketReader, BucketReader>()
+                .AddSingleton<IBucketReader, S3BucketReader>()
+                .AddSingleton<IBucketWriter, S3BucketWriter>()
                 .AddSingleton<IThumbReorganiser, NonOrganisingReorganiser>()
                 .AddTransient<IAssetPathGenerator, ConfigDrivenAssetPathGenerator>()
                 .AddScoped<AccessChecker>()
